dev-requirements.txt
flake8    # for the travis build
ckanext-scheming # imported in tests
pytest-randomly
